Remove the "browser" env in ESLint config (#13)
We should not assume that our code will be run in a browser.  Instead,
the code is written to be compatible with both the browser and the
Node.js environments.

This also implies that we should not use any feature that’s unique any
of these environments, such as the window or the process object.

See https://eslint.org/docs/user-guide/configuring for more information.